Faster and better Main.jar
Well I made it easier in a way and better. And even faster probably by a lot. Just read the comments! ;)

To be honest, I didn't test it out, so you can expect errors. And I did import the whole code into a project of my own into Eclipse, so that too. I'm pretty sure it works fine now! :D